pyproject.toml
[project]
name = "claims"
version = "0.0.1"
dependencies = [
    "django>=5.2.8",
    "pytesseract>=0.3.10",
    "pillow>=10.0.0",
    "psycopg2-binary>=2.9.9",
    "pytesseract>=0.3.13",
]
